moved styles in separate file
