+++ src/components/LoadingSpinner/index.tsx
@@ -1,0 +1,11 @@
+import { StyledStack, StyledCircularProgress } from "./LoadingSpinnerStyles";
+
+const LoadingSpinner = () => {
+  return (
+    <StyledStack direction="row">
+      <StyledCircularProgress />
+    </StyledStack>
+  );
+};
+
+export default LoadingSpinner;
